+++ .idea/misc.xml
@@ -1,5 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="..\:/Users/ssiss/Desktop/barcode_scanner/barcode-app-workspace/app/src/main/res/layout/activity_goodsinfo.xml" value="0.33" />
+        <entry key="..\:/Users/ssiss/Desktop/barcode_scanner/barcode-app-workspace/app/src/main/res/layout/activity_main.xml" value="0.36614583333333334" />
+        <entry key="..\:/Users/ssiss/Desktop/barcode_scanner/barcode-app-workspace/app/src/main/res/layout/activity_scan.xml" value="0.36614583333333334" />
+      </map>
+    </option>
+  </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
