+++ app/src/main/java/com/SE_project/barcode_scanner/MainActivity.java
@@ -9,6 +9,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
+import android.widget.Toast;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -18,15 +19,31 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
         setTitle("바코드 스캐너");
 
-        ImageView ivCamera = findViewById(R.id.ivCamera);
-        ImageView ivGallery = findViewById(R.id.ivGallery);
+        Button SuccessScan = findViewById(R.id.btnSuccess);
+        Button FailScan = findViewById(R.id.btnFail);
 
-        ivCamera.setOnClickListener(new View.OnClickListener() {
+        SuccessScan.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent i = new Intent(getApplicationContext(), ScanActivity.class);
+                Intent i = new Intent(getApplicationContext(), GoodsInfoActivity.class);
                 startActivity(i);
             }
         });
+
+        FailScan.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Toast.makeText(getApplicationContext(), "인식 실패", Toast.LENGTH_SHORT).show();
+            }
+        });
+
+        ImageView ivGallery = findViewById(R.id.ivGallery);
+        ivGallery.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+            }
+        });
+
     }
 }
