+++ app/src/main/res/layout/activity_main.xml
@@ -7,24 +7,59 @@
     android:orientation="vertical"
     tools:context=".MainActivity">
 
-    <ImageView
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:layout_weight="1"
-        android:layout_margin="20dp"
-        android:src="@drawable/camera"
-        android:id="@+id/ivCamera"
-        />
+        android:layout_weight="11"
+        android:gravity="center"
+        android:orientation="vertical">
+        <ImageView
+            android:layout_width="350dp"
+            android:layout_height="400dp"
+            android:background="@color/black"
+            />
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="5dp"
+            android:text="(카메라 화면)"
+            android:textSize="20sp"
+            />
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal"
+            >
+            <Button
+                android:layout_width="0dp"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:layout_margin="10dp"
+                android:text="인식성공"
+                android:id="@+id/btnSuccess"
+                />
+            <Button
+                android:layout_width="0dp"
+                android:layout_height="match_parent"
+                android:layout_weight="1"
+                android:layout_margin="10dp"
+                android:text="인식실패"
+                android:id="@+id/btnFail"
+                />
+        </LinearLayout>
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="(테스트를 위한 버튼)"
+            />
+    </LinearLayout>
     <ImageView
-        android:layout_width="match_parent"
+        android:layout_width="50dp"
         android:layout_height="0dp"
         android:layout_weight="1"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
-        android:layout_marginBottom="20dp"
         android:src="@drawable/gallery"
+        android:layout_gravity="right"
+        android:layout_margin="10dp"
         android:id="@+id/ivGallery"
         />
-
-
 </LinearLayout>
